extract function for checking reserved prefixes
